requirements.txt
discord.py[voice]
python-dotenv
PyNaCl
youtube_dl